dashboard/requirements.txt
streamlit
pandas
plotly
python-dotenv
langchain
langchain-groq
